pipeline fix

+++ config/system.php
@@ -38,7 +38,7 @@ return [
     |
     */
     'operation_mode' => env('APP_OPERATION_MODE', 'standalone'),
-    'standalone_jwt_ttl_minutes' => env('STANDALONE_JWT_TTL_MINUTES', 60),
+    'standalone_jwt_ttl_minutes' => (int)env('STANDALONE_JWT_TTL_MINUTES', 60),
 
     /*
     |--------------------------------------------------------------------------
